+++ include/mementar/lz/BitFileGetter.h
@@ -7,12 +7,12 @@
 class BitFileGetter
 {
 public:
-  BitFileGetter(size_t size_1 = 0, size_t size_2 = 0, size_t size_3 = 0, size_t size_4 = 0);
+  BitFileGetter(uint32_t size_1 = 0, uint32_t size_2 = 0, uint32_t size_3 = 0, uint32_t size_4 = 0);
 
-  void setSize1(size_t size_1) { type_1_size_ = size_1; }
-  void setSize2(size_t size_2) { type_2_size_ = size_2; }
-  void setSize3(size_t size_3) { type_3_size_ = size_3; }
-  void setSize4(size_t size_4) { type_4_size_ = size_4; }
+  void setSize1(uint32_t size_1) { type_1_size_ = size_1; }
+  void setSize2(uint32_t size_2) { type_2_size_ = size_2; }
+  void setSize3(uint32_t size_3) { type_3_size_ = size_3; }
+  void setSize4(uint32_t size_4) { type_4_size_ = size_4; }
 
   uint32_t getType1();
   uint32_t getType2();
@@ -29,12 +29,12 @@ public:
 private:
   std::vector<char> data_;
   size_t major_index_;
-  size_t minor_index_;
+  uint8_t minor_index_;
 
-  size_t type_1_size_;
-  size_t type_2_size_;
-  size_t type_3_size_;
-  size_t type_4_size_;
+  uint32_t type_1_size_;
+  uint32_t type_2_size_;
+  uint32_t type_3_size_;
+  uint32_t type_4_size_;
 };
 
 #endif // BITFILEGETTER_H
